requirements.txt
# Core AI and ML libraries
crewai==0.148.0
langchain==0.3.26
langchain-community==0.3.26
langchain-openai==0.3.26
langchain-anthropic==0.3.16
langchain-groq==0.3.5
langchain-google-genai==2.0.4
langchain-together==0.3.0
langchain-mistralai==0.2.11
langchain-ollama==0.3.5
langchain-core==0.3.69
anthropic==0.57.1
openai==1.97.0
groq==0.30.0
google-generativeai==0.8.0

# Official MCP SDK
mcp==1.12.0

# Data processing and analysis
pandas
numpy
scipy
scikit-learn
# State-of-the-art ML models (2025 upgrades)
xgboost==2.1.3
lightgbm==4.5.0
catboost==1.2.7
gliner==0.2.8

# Analytics and performance libraries
empyrical==0.5.5
quantstats
pyfolio
statsmodels

yfinance
alpha-vantage

# Plotting and visualization
matplotlib
seaborn
plotly

# Utilities and performance
joblib==1.5.1

# Web and API
requests==2.31.0
aiohttp==3.12.14
fastapi==0.116.1
uvicorn==0.35.0
httpx==0.26.0

# Social media and scraping
praw==7.8.1
tweepy==4.16.0

# Database
sqlalchemy==2.0.41
redis==5.3.0
pymongo==4.13.2
psycopg2-binary==2.9.10

# Monitoring and logging
structlog==23.3.0
prometheus-client==0.22.1
psutil==5.9.8
sentry-sdk[fastapi]==2.17.0

# Configuration and utilities
python-dotenv==1.1.1
pydantic==2.11.7
click==8.2.1
pyyaml==6.0.2
watchdog==3.0.0
tqdm==4.67.1

# Communication and messaging
asyncio-mqtt==0.16.2
websockets

# Trading and financial data
ccxt
polygon-api-client
alpaca-trade-api
vectorbt
QuantLib-Python

# Advanced analytics and ML
hmmlearn==0.3.3
ruptures==1.1.9
changepoint==0.3.2
cvxpy==1.7.0
PyPortfolioOpt==1.5.6

# Reinforcement learning
gym==0.26.2
stable-baselines3==2.6.0

# Deep learning
tensorflow==2.19.0
tf-keras==2.19.0

# NLP and transformers
transformers==4.53.2
torch==2.7.1
sentence-transformers==2.7.0
spacy==3.8.7

# Message queue
kafka-python==2.2.15

# Additional utilities
python-multipart==0.0.20
aiofiles==23.2.1
textblob
slowapi
prometheus-fastapi-instrumentator
passlib[bcrypt]
opencv-python
pytrends
feedparser
arch
asyncpraw 